requirements.txt
# UI
streamlit==1.39.0

# LLM / APIs
openai>=1.54.0
python-dotenv==1.0.1

# MCP
mcp[cli]>=1.27.0
nest_asyncio==1.6.0

# Data
pandas==2.2.2
numpy==1.26.4
pydantic>=2.8.0

# RAG opcional
langchain>=0.3.7
langchain-community>=0.3.5
langchain-chroma>=0.1.4
langchain-huggingface>=0.1.0
chromadb>=0.5.18
sentence-transformers>=3.2.1
tiktoken>=0.8.0
pypdf>=5.0.0
docx2txt>=0.8
unstructured[md]>=0.15.0

# Utilidades
tqdm>=4.66.0
